chore: Move Telemetry `SerializerType` enum from PHPUnit to OTLP
