+++ src/bridge/phpunit/telemetry/tests/Flow/Bridge/PHPUnit/Telemetry/Tests/Unit/TelemetryFactoryTest.php
@@ -9,13 +9,13 @@ use Flow\Bridge\PHPUnit\Telemetry\CurlTransportConfig;
 use Flow\Bridge\PHPUnit\Telemetry\ErrorLogHandlerConfig;
 use Flow\Bridge\PHPUnit\Telemetry\GrpcTransportConfig;
 use Flow\Bridge\PHPUnit\Telemetry\NullErrorHandlerConfig;
-use Flow\Bridge\PHPUnit\Telemetry\SerializerType;
 use Flow\Bridge\PHPUnit\Telemetry\StreamErrorHandlerConfig;
 use Flow\Bridge\PHPUnit\Telemetry\StreamTransportConfig;
 use Flow\Bridge\PHPUnit\Telemetry\SyslogErrorHandlerConfig;
 use Flow\Bridge\PHPUnit\Telemetry\TelemetryFactory;
 use Flow\Bridge\PHPUnit\Telemetry\Tests\Mother\ConfigurationMother;
 use Flow\Bridge\PHPUnit\Telemetry\UdpSyslogErrorHandlerConfig;
+use Flow\Bridge\Telemetry\OTLP\Serializer\SerializerType;
 use Flow\Telemetry\ErrorHandler\ErrorLogMessageType;
 use Flow\Telemetry\ErrorHandler\SyslogFacility;
 use Flow\Telemetry\ErrorHandler\SyslogSeverity;
